Implement dual-kernel baseline with warp-shuffle stencils.
* Implement `magnetic_kernel` and `electric_kernel` for H/E field updates.
* Establish $N \times 3$ interleaved float layout with stride-3 indexing.
* Magnetic Kernel (Faraday Step): Implemented magnetic_kernel to update the H-field based on the curl of E. It includes placeholders for the subsequent LLG RK4 integration and M-H coupling.
* Electric Kernel (Ampere Step): Implemented `electric_kernel` to update the E-field based on the curl of H, preparing the stage for future ABC (Absorbing Boundary Condition) logic.

+++ gpu_tools/llg_kernel.cu
@@ -2,9 +2,9 @@
 
 // Helper: Cross Product (Register-only)
 __device__ __forceinline__ void cross_product(
-    double ax, double ay, double az,
-    double bx, double by, double bz,
-    double &rx, double &ry, double &rz) 
+    float ax, float ay, float az,
+    float bx, float by, float bz,
+    float &rx, float &ry, float &rz) 
 {
     rx = ay * bz - az * by;
     ry = az * bx - ax * bz;
@@ -13,17 +13,17 @@ __device__ __forceinline__ void cross_product(
 
 // Helper: LLG Derivative Logic
 __device__ __forceinline__ void compute_llg_derivative(
-    double Mx, double My, double Mz,      
-    double Hx, double Hy, double Hz,      
-    double neg_gamma_LL, double neg_coeff_damp,  // Constants
-    double &kx, double &ky, double &kz)     // Output 
+    float Mx, float My, float Mz,      
+    float Hx, float Hy, float Hz,      
+    float neg_gamma_LL, float neg_coeff_damp,  // Constants
+    float &kx, float &ky, float &kz)     // Output 
 {
     // M x H
-    double c1x, c1y, c1z;
+    float c1x, c1y, c1z;
     cross_product(Mx, My, Mz, Hx, Hy, Hz, c1x, c1y, c1z);
 
     // M x (M x H)
-    double c2x, c2y, c2z;
+    float c2x, c2y, c2z;
     cross_product(Mx, My, Mz, c1x, c1y, c1z, c2x, c2y, c2z);
 
     kx = neg_gamma_LL * c1x + neg_coeff_damp * c2x;
@@ -33,13 +33,13 @@ __device__ __forceinline__ void compute_llg_derivative(
 
 // Main Kernel
 __global__ void LLG_RK4_kernel(
-    const double* __restrict__ M_curr,  // Real
-    const double* __restrict__ H_curr,  // Complex, only real part (double) needed
-    double* __restrict__ M_next,       
+    const float* __restrict__ M_curr,  // Real
+    const float* __restrict__ H_curr,  // Complex, only real part (float) needed
+    float* __restrict__ M_next,       
     int material_size,                            
-    double dt,                         
-    double neg_gamma_LL,  // -gamma_LL (negative)
-    double neg_coeff_damp // -(gamma * alpha * factor) / M0
+    float dt,                         
+    float neg_gamma_LL,  // -gamma_LL (negative)
+    float neg_coeff_damp // -(gamma * alpha * factor) / M0
 ) 
 {
     int idx = blockIdx.x * blockDim.x + threadIdx.x;
@@ -47,19 +47,19 @@ __global__ void LLG_RK4_kernel(
 
     // Load initial state to registers
     int ptr_m = idx * 3; 
-    double Mx = M_curr[ptr_m];
-    double My = M_curr[ptr_m + 1];
-    double Mz = M_curr[ptr_m + 2];
+    float Mx = M_curr[ptr_m];
+    float My = M_curr[ptr_m + 1];
+    float Mz = M_curr[ptr_m + 2];
 
     int ptr_h = idx * 6; // complex
-    double Hx = H_curr[ptr_h];
-    double Hy = H_curr[ptr_h + 2]; // next real part
-    double Hz = H_curr[ptr_h + 4];
+    float Hx = H_curr[ptr_h];
+    float Hy = H_curr[ptr_h + 2]; // next real part
+    float Hz = H_curr[ptr_h + 4];
 
     // RK4 Accumulators
-    double kx, ky, kz;
-    double sum_x = 0.0, sum_y = 0.0, sum_z = 0.0;
-    double half_dt = 0.5*dt;
+    float kx, ky, kz;
+    float sum_x = 0.0, sum_y = 0.0, sum_z = 0.0;
+    float half_dt = 0.5*dt;
 
     // Step 1
     compute_llg_derivative(Mx, My, Mz, Hx, Hy, Hz, neg_gamma_LL, neg_coeff_damp, kx, ky, kz);
@@ -81,7 +81,7 @@ __global__ void LLG_RK4_kernel(
     sum_x += kx; sum_y += ky; sum_z += kz;
 
     // Final Update
-    double dt_div_6 = dt / 6.0;
+    float dt_div_6 = dt / 6.0;
     M_next[ptr_m]     = Mx + dt_div_6 * sum_x;
     M_next[ptr_m + 1] = My + dt_div_6 * sum_y;
     M_next[ptr_m + 2] = Mz + dt_div_6 * sum_z;
